added xml-rpc based remote server mode for nfsim
added a preliminary implementation of an xml-rpc based service for nfsim. The service loads up a bng-xml
file and offers three methods: simulate, equilibrate and printing system status to stdout. The server is based on the
xml-rpc library.

Also added a cmake-based build chain to facilitate debugging and adding libraries.